+++ cg/apps/lims/__init__.py
@@ -1,3 +1,3 @@
 from .api import LimsAPI
 from .orderform import parse_orderform
-from .limsjson import parse_json
+from .limsjson import parse_json_order