+++ cg/cli/upload/delivery_report.py
@@ -1,5 +1,4 @@
 """Code for uploading delivery report from the CLI"""
-
 import logging
 import sys
 from pathlib import Path