+++ tests/apps/mip/conftest.py
@@ -95,8 +95,8 @@ def cg_config():
 
 @pytest.fixture
 def link_family():
-    """mock family name"""
-    return "family"
+    """mock case name"""
+    return "case"
 
 
 @pytest.fixture