+++ tests/cli/test_cli_add_relationship.py
@@ -5,11 +5,11 @@ from cg.store import Store
 
 def test_add_relationship_required(invoke_cli, disk_store: Store, helpers):
     """Test to add a relationship using only the required arguments"""
-    # GIVEN a database with a sample and an family
+    # GIVEN a database with a sample and an case
     sample = helpers.add_sample(disk_store)
     sample_id = sample.internal_id
-    family = helpers.add_family(disk_store)
-    family_id = family.internal_id
+    case = helpers.add_case(disk_store)
+    case_id = case.internal_id
     status = "affected"
 
     # WHEN adding a relationship
@@ -21,7 +21,7 @@ def test_add_relationship_required(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,
@@ -31,7 +31,7 @@ def test_add_relationship_required(invoke_cli, disk_store: Store, helpers):
     # THEN then it should be added
     assert result.exit_code == 0
     assert disk_store.FamilySample.query.count() == 1
-    assert disk_store.FamilySample.query.first().family.internal_id == family_id
+    assert disk_store.FamilySample.query.first().family.internal_id == case_id
     assert disk_store.FamilySample.query.first().sample.internal_id == sample_id
 
 
@@ -41,8 +41,8 @@ def test_add_relationship_bad_sample(invoke_cli, disk_store: Store, helpers):
 
     # WHEN adding a relationship
     db_uri = disk_store.uri
-    family = helpers.add_family(disk_store)
-    family_id = family.internal_id
+    case = helpers.add_case(disk_store)
+    case_id = case.internal_id
     sample_id = "dummy_sample"
     status = "affected"
     result = invoke_cli(
@@ -51,7 +51,7 @@ def test_add_relationship_bad_sample(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,
@@ -64,12 +64,12 @@ def test_add_relationship_bad_sample(invoke_cli, disk_store: Store, helpers):
 
 
 def test_add_relationship_bad_family(invoke_cli, disk_store: Store, helpers):
-    """Test to add a relationship using a non-existing family"""
+    """Test to add a relationship using a non-existing case"""
     # GIVEN a database with a sample
 
     # WHEN adding a relationship
     db_uri = disk_store.uri
-    family_id = "dummy_family"
+    case_id = "dummy_family"
     sample = helpers.add_sample(disk_store)
     sample_id = sample.internal_id
 
@@ -80,29 +80,29 @@ def test_add_relationship_bad_family(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,
         ]
     )
 
-    # THEN then it should complain in missing family instead of adding a relationship
+    # THEN then it should complain in missing case instead of adding a relationship
     assert result.exit_code == 1
     assert disk_store.FamilySample.query.count() == 0
 
 
 def test_add_relationship_bad_status(invoke_cli, disk_store: Store, helpers):
     """Test that the added relationship get the status we send in"""
-    # GIVEN a database with a sample and an family
+    # GIVEN a database with a sample and an case
 
     # WHEN adding a relationship
     db_uri = disk_store.uri
 
     sample = helpers.add_sample(disk_store)
     sample_id = sample.internal_id
-    family = helpers.add_family(disk_store)
-    family_id = family.internal_id
+    case = helpers.add_case(disk_store)
+    case_id = case.internal_id
     status = "dummy_status"
 
     result = invoke_cli(
@@ -111,7 +111,7 @@ def test_add_relationship_bad_status(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,
@@ -125,13 +125,13 @@ def test_add_relationship_bad_status(invoke_cli, disk_store: Store, helpers):
 
 def test_add_relationship_mother(invoke_cli, disk_store: Store, helpers):
     """Test to add a relationship with a mother"""
-    # GIVEN a database with a sample and an family
+    # GIVEN a database with a sample and an case
     sample = helpers.add_sample(disk_store)
     sample_id = sample.internal_id
     mother = helpers.add_sample(disk_store, sample_id="mother")
     mother_id = mother.internal_id
-    family = helpers.add_family(disk_store)
-    family_id = family.internal_id
+    case = helpers.add_case(disk_store)
+    case_id = case.internal_id
     status = "affected"
 
     # WHEN adding a relationship
@@ -143,7 +143,7 @@ def test_add_relationship_mother(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,
@@ -160,12 +160,12 @@ def test_add_relationship_mother(invoke_cli, disk_store: Store, helpers):
 
 def test_add_relationship_bad_mother(invoke_cli, disk_store: Store, helpers):
     """Test to add a relationship using a non-existing mother"""
-    # GIVEN a database with a sample and an family
+    # GIVEN a database with a sample and an case
     sample = helpers.add_sample(disk_store)
     sample_id = sample.internal_id
     mother_id = "dummy_mother"
-    family = helpers.add_family(disk_store)
-    family_id = family.internal_id
+    case = helpers.add_case(disk_store)
+    case_id = case.internal_id
     status = "affected"
 
     # WHEN adding a relationship
@@ -177,7 +177,7 @@ def test_add_relationship_bad_mother(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,
@@ -193,15 +193,15 @@ def test_add_relationship_bad_mother(invoke_cli, disk_store: Store, helpers):
 
 def test_add_relationship_father(invoke_cli, disk_store: Store, helpers):
     """Test to add a relationship using a father"""
-    # GIVEN a database with a sample and an family
+    # GIVEN a database with a sample and an case
     sample = helpers.add_sample(disk_store)
     sample_id = sample.internal_id
 
     father = helpers.add_sample(disk_store, sample_id="father", gender="male")
     father_id = father.internal_id
 
-    family = helpers.add_family(disk_store)
-    family_id = family.internal_id
+    case = helpers.add_case(disk_store)
+    case_id = case.internal_id
 
     status = "affected"
 
@@ -214,7 +214,7 @@ def test_add_relationship_father(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,
@@ -231,13 +231,13 @@ def test_add_relationship_father(invoke_cli, disk_store: Store, helpers):
 
 def test_add_relationship_bad_father(invoke_cli, disk_store: Store, helpers):
     """Test to add a relationship using a non-existing father"""
-    # GIVEN a database with a sample and an family
+    # GIVEN a database with a sample and an case
     sample = helpers.add_sample(disk_store)
     sample_id = sample.internal_id
 
     father_id = "bad_father"
-    family = helpers.add_family(disk_store)
-    family_id = family.internal_id
+    case = helpers.add_case(disk_store)
+    case_id = case.internal_id
 
     status = "affected"
 
@@ -250,7 +250,7 @@ def test_add_relationship_bad_father(invoke_cli, disk_store: Store, helpers):
             db_uri,
             "add",
             "relationship",
-            family_id,
+            case_id,
             sample_id,
             "-s",
             status,