+++ CHANGELOG.md
@@ -19,6 +19,34 @@ Try to use the following format:
 ### Added
 - Support for creating delivery reports for analyses that are not the latest ones
 
+## [18.10.0]
+
+### Added
+ - Support for delivery type in the Order Portal
+
+## [18.9.1]
+
+### Added
+- cg upload vogue bioinfo-all uploads both BALSAMIC as well.
+
+## [18.9.0]
+
+### Added
+
+- Add functionality to upload balsamic analyses to scout
+
+## [18.8.0]
+
+### Added
+- cg workflow microsalt upload-analysis-vogue [case_id] to upload the latest analysis from specific case
+- cg workflow microsalt upload-latest-analyses-vogue to upload all latest analyses what haven't been uploaded
+
+## [18.7.2]
+
+### Changed
+
+- Skip deliver fastq files when delivering balsamic analysis
+
 ## [18.7.1]
 
 ### Fixed
@@ -27,7 +55,8 @@ Try to use the following format:
 
 ## [18.7.0]
 
-### Added customer name in order tickets
+### Added
+- Added customer name in order tickets
 
 ## [18.6.1]
 
@@ -35,7 +64,9 @@ Try to use the following format:
 
 - Fix bug with name clash for created case when submitting RML-orders via Orderportal
 
-## 18.6.0
+
+## [18.6.0]
+
 
 ### Added